It's generally a good idea to fix dependencies to a version.
Otherwise, they can make breaking changes and break your program. This
is less of an issue with binary packages because you check in Cargo.lock
but is still a good idea.

The rule is:

If the library is version 0.0.z, use 0.0.z (the full version).
If the library is version 0.y.z, use 0.y (allow patch releases).
If the library is version x.y.z, use x (no breaking changes) or x.y (no new features).